+++ tune n-edo.qml
@@ -4,7 +4,7 @@ import QtQuick.Controls.Styles 1.3
 import MuseScore 3.0
 
 MuseScore {
-      version: "2.2.6"
+      version: "2.2.8"
       description: "Retune selection to any EDO temperament, or whole score if nothing selected."
       menuPath: "Plugins.n-EDO.Tune"
 
@@ -380,9 +380,9 @@ MuseScore {
                 // scan edo & tuning center first. key signature parsing is dependant on edo used.
                 for (var i = 0; i < cursor.segment.annotations.length; i++) {
                   var annotation = cursor.segment.annotations[i];
-                  console.log("found annotation type: " + annotation.subtypeName());
-                  if ((annotation.subtypeName() == 'Staff' && Math.floor(annotation.track / 4) == staff) ||
-                      (annotation.subtypeName() == 'System')) {
+                  console.log("found annotation type: " + annotation.name);
+                  if ((annotation.name == 'StaffText' && Math.floor(annotation.track / 4) == staff) ||
+                      (annotation.name == 'SystemText')) {
                     var text = removeFormattingCode(annotation.text);
                     if (text.toLowerCase().trim().endsWith('edo')) {
                       var edo = parseInt(text.substring(0, text.length - 3));
@@ -412,9 +412,9 @@ MuseScore {
                 // Check for StaffText key signature changes, then update staffKeySigHistory
                 for (var i = 0; i < cursor.segment.annotations.length; i++) {
                   var annotation = cursor.segment.annotations[i];
-                  console.log("found annotation type: " + annotation.subtypeName());
-                  if ((annotation.subtypeName() == 'Staff' && Math.floor(annotation.track / 4) == staff) ||
-                      (annotation.subtypeName() == 'System')) {
+                  console.log("found annotation type: " + annotation.name);
+                  if ((annotation.name == 'StaffText' && Math.floor(annotation.track / 4) == staff) ||
+                      (annotation.name == 'SystemText')) {
                     var text = removeFormattingCode(annotation.text);
                     var mostRecentEDO = staffEDOHistory.length !== 0 ? staffEDOHistory[staffEDOHistory.length - 1].edo : null;
                     if (!mostRecentEDO)